celebal-internship-tasks/Week-05/azure_airflow_project/requirements.txt
pandas
sqlalchemy
pyarrow
fastavro
azure-storage-blob
python-dotenv
apache-airflow
